+++ test/spec/minireel/ad_manager.controller.ut.js
@@ -723,7 +723,8 @@
                                     skip: 6
                                 },
                                 display: {
-                                    waterfall: 'cinema6'
+                                    waterfall: 'cinema6',
+                                    enabled: false
                                 }
                             },
                             data: {
@@ -741,7 +742,8 @@
                                 skip: false
                             },
                             display: {
-                                waterfall: 'cinema6'
+                                waterfall: 'cinema6',
+                                enabled: true
                             }
                         };
 
@@ -758,7 +760,8 @@
                                     skip: false
                                 },
                                 display: {
-                                    waterfall: 'cinema6'
+                                    waterfall: 'cinema6',
+                                    enabled: true
                                 }
                             },
                             data: {
@@ -771,7 +774,8 @@
                                         skip: false
                                     },
                                     display: {
-                                        waterfall: 'cinema6'
+                                        waterfall: 'cinema6',
+                                        enabled: true
                                     }
                                 }
                             }
@@ -806,7 +810,8 @@
                                     skip: 6
                                 },
                                 display: {
-                                    waterfall: 'cinema6'
+                                    waterfall: 'cinema6',
+                                    enabled: true
                                 }
                             };
 
@@ -823,7 +828,8 @@
                                         skip: 6
                                     },
                                     display: {
-                                        waterfall: 'cinema6'
+                                        waterfall: 'cinema6',
+                                        enabled: true
                                     }
                                 },
                                 data: [minireel]
@@ -846,7 +852,8 @@
                                                 skip: 6
                                             },
                                             display: {
-                                                waterfall: 'cinema6'
+                                                waterfall: 'cinema6',
+                                                enabled: true
                                             }
                                         }
                                     }
@@ -862,7 +869,8 @@
                                                 skip: 6
                                             },
                                             display: {
-                                                waterfall: 'cinema6'
+                                                waterfall: 'cinema6',
+                                                enabled: true
                                             }
                                         }
                                     }
@@ -882,7 +890,8 @@
                                         skip: 6
                                     },
                                     display: {
-                                        waterfall: 'cinema6'
+                                        waterfall: 'cinema6',
+                                        enabled: true
                                     }
                                 },
                                 data: minireels
@@ -901,7 +910,8 @@
                                                 skip: true
                                             },
                                             display: {
-                                                waterfall: 'cinema6'
+                                                waterfall: 'cinema6',
+                                                enabled: true
                                             }
                                         }
                                     }
@@ -917,7 +927,8 @@
                                                 skip: 6
                                             },
                                             display: {
-                                                waterfall: 'publisher'
+                                                waterfall: 'publisher',
+                                                enabled: false
                                             }
                                         }
                                     }
@@ -937,7 +948,8 @@
                                         skip: undefined
                                     },
                                     display: {
-                                        waterfall: undefined
+                                        waterfall: undefined,
+                                        enabled: undefined
                                     }
                                 },
                                 data: minireels
@@ -956,7 +968,8 @@
                                                 skip: 6
                                             },
                                             display: {
-                                                waterfall: 'cinema6'
+                                                waterfall: 'cinema6',
+                                                enabled: false
                                             }
                                         }
                                     }
@@ -972,7 +985,8 @@
                                                 skip: 6
                                             },
                                             display: {
-                                                waterfall: 'publisher'
+                                                waterfall: 'publisher',
+                                                enabled: false
                                             }
                                         }
                                     }
@@ -992,7 +1006,8 @@
                                         skip: 6
                                     },
                                     display: {
-                                        waterfall: undefined
+                                        waterfall: undefined,
+                                        enabled: false
                                     }
                                 },
                                 data: minireels
@@ -1002,53 +1017,55 @@
                 });
 
                 describe('useDefaultSettings()', function() {
-                    it('should go through all selected minireels and delete static ad cards and adConfig blocks', function() {
-                        var onAffirm, onCancel,
-                            minireels = [
-                                {
-                                    id: 'e-1',
-                                    data: {
-                                        adConfig: {},
-                                        deck: [
-                                            {
-                                                ad: true
-                                            }
-                                        ]
-                                    },
-                                    save: jasmine.createSpy('minireel.save()')
-                                },
-                                {
-                                    id: 'e-2',
-                                    data: {
-                                        adConfig: {},
-                                        deck: [
-                                            {
-                                                ad: true
-                                            },
-                                            {
-                                                ad: true
-                                            }
-                                        ]
-                                    },
-                                    save: jasmine.createSpy('minireel.save()')
-                                },
-                                {
-                                    id: 'e-3',
-                                    data: {
-                                        adConfig: {},
-                                        deck: []
-                                    },
-                                    save: jasmine.createSpy('minireel.save()')
-                                }
-                            ];
+                    var onAffirm, onCancel, minireels;
 
+                    beforeEach(function() {
+                        minireels = [
+                            {
+                                id: 'e-1',
+                                data: {
+                                    adConfig: {},
+                                    deck: [
+                                        {
+                                            ad: true
+                                        }
+                                    ]
+                                },
+                                save: jasmine.createSpy('minireel.save()')
+                            },
+                            {
+                                id: 'e-2',
+                                data: {
+                                    adConfig: {},
+                                    deck: [
+                                        {
+                                            ad: true
+                                        },
+                                        {
+                                            ad: true
+                                        }
+                                    ]
+                                },
+                                save: jasmine.createSpy('minireel.save()')
+                            },
+                            {
+                                id: 'e-3',
+                                data: {
+                                    adConfig: {},
+                                    deck: []
+                                },
+                                save: jasmine.createSpy('minireel.save()')
+                            }
+                        ];
 
                         spyOn(ConfirmDialogService, 'display');
                         AdManagerCtrl.useDefaultSettings(minireels);
 
                         onAffirm = ConfirmDialogService.display.calls.mostRecent().args[0].onAffirm;
                         onCancel = ConfirmDialogService.display.calls.mostRecent().args[0].onCancel;
+                    });
 
+                    it('should go through all selected minireels and delete static ad cards and adConfig blocks', function() {
                         expect(ConfirmDialogService.display).toHaveBeenCalled();
 
                         onAffirm();
@@ -1058,6 +1075,40 @@
                             expect(minireel.data.deck.length).toBe(0);
                             expect(minireel.save).toHaveBeenCalled();
                         });
+                    });
+
+                    it('should remove displayAd modules if no Org adConfig is defined', function() {
+                        spyOn(MiniReelService, 'enableDisplayAds').and.callThrough();
+                        spyOn(MiniReelService, 'disableDisplayAds').and.callThrough();
+
+                        onAffirm();
+
+                        expect(MiniReelService.enableDisplayAds).not.toHaveBeenCalled();
+                        expect(MiniReelService.disableDisplayAds).toHaveBeenCalled();
+                    });
+
+                    it('should add displayAd modules based on Org adConfig', function() {
+                        PortalCtrl.model.org.adConfig = { display: { enabled: true }};
+
+                        spyOn(MiniReelService, 'enableDisplayAds').and.callThrough();
+                        spyOn(MiniReelService, 'disableDisplayAds').and.callThrough();
+
+                        onAffirm();
+
+                        expect(MiniReelService.enableDisplayAds).toHaveBeenCalled();
+                        expect(MiniReelService.disableDisplayAds).not.toHaveBeenCalled();
+                    });
+
+                    it('should remove displayAd modules based on Org adConfig', function() {
+                        PortalCtrl.model.org.adConfig = { display: { enabled: false }};
+
+                        spyOn(MiniReelService, 'enableDisplayAds').and.callThrough();
+                        spyOn(MiniReelService, 'disableDisplayAds').and.callThrough();
+
+                        onAffirm();
+
+                        expect(MiniReelService.enableDisplayAds).not.toHaveBeenCalled();
+                        expect(MiniReelService.disableDisplayAds).toHaveBeenCalled();
                     });
                 });